--- .pypirc
@@ -1,12 +1,0 @@
-[distutils]
-index-servers =
-   pypi
-   testpypi
-
-[pypi]
-repository = https://upload.pypi.org/legacy/
-username = jobala
-
-[testpypi]
-repository = https://test.pypi.org/legacy/
-username = jobala
