+++ README.md
@@ -1,5 +1,4 @@
-Microsoft Graph Python Client Library
-=====================================
+## Microsoft Graph Python Client Library
 
 The Microsoft Graph Python client library is a lightweight wrapper around
 the Microsoft Graph API.
