+++ dev_requirements.txt
@@ -1,3 +1,4 @@
 coverage==5.0.3
 pylint==2.4.4
 responses==0.10.12
+flit==2.2.0
