+++ msgraphcore/constants.py
