+++ msgraphcore/middleware/_base_auth.py
