+++ tests/unit/test_middleware_options.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
 
-from src.middleware.options.auth_middleware_options import AuthMiddlewareOptions
+from msgraphcore.middleware.options.auth_middleware_options import AuthMiddlewareOptions
 
 
 class TestMiddlewareOptions(TestCase):
